feat: Add Rating and Review functionality with DTOs and Controller updates
- Implemented RatingDTO to handle ratings and reviews for spaces.
- Updated SpaceDTO to include average rating and review count.
- Enhanced AdminDashboardController to display ratings in the UI.
- Added CSS styles for image upload and rating display.
- Created RatingController with endpoints for managing ratings.
- Updated SpaceController to support creating spaces with images.
- Enhanced RatingService and its implementation for new rating features.
- Added repository methods for fetching ratings and calculating averages.
- Updated EmailService to handle new reservation statuses.
- Added tests for SpaceController to validate new features.

+++ src/main/java/finalprojectprogramming/project/models/Space.java
@@ -62,7 +62,7 @@ public class Space {
 
     @Builder.Default
     @Column(name = "average_rating")
-    private Float averageRating = 0f;
+    private Double averageRating = 0.0;
 
     @Column(name = "created_at")
     private LocalDateTime createdAt;
